+++ custom_components/welcomeeye_local/const.py
@@ -1,2 +1,2 @@
 DOMAIN = 'welcomeeye_local'
-DEFAULT_NAME = 'WelcomeEye Connect 2'
+DEFAULT_NAME = 'Philips WelcomeEye'